Add support for Quasar initialization options.
Using ESM vs UMD for Quasar means plugins are no longer used
automatically or by default. Applications must be initialized with
Quasar plugins it uses. This is currently not automatic and requires
top-level application knowledge of all sub-modules in use and what
plugins they require. The `quasarOptions` object added here can be used
for this purpose and for other Quasar options.